Extract current input mode into its own library
Removes a transitive dependency on the entire plrctrl from options.
