+++ Source/controls/controller_motion.cpp
@@ -3,6 +3,7 @@
 #include <cmath>
 
 #include "control.h"
+#include "controls/control_mode.hpp"
 #include "controls/controller.h"
 #ifndef USE_SDL1
 #include "controls/devices/game_controller.h"
