+++ Source/controls/game_controls.cpp
@@ -2,6 +2,7 @@
 
 #include <cstdint>
 
+#include "controls/control_mode.hpp"
 #include "controls/controller_motion.h"
 #ifndef USE_SDL1
 #include "controls/devices/game_controller.h"
