+++ Source/controls/game_controls.h
@@ -8,13 +8,6 @@
 
 namespace devilution {
 
-enum class GamepadLayout : uint8_t {
-	Generic,
-	Nintendo,
-	PlayStation,
-	Xbox,
-};
-
 enum GameActionType : uint8_t {
 	GameActionType_NONE,
 	GameActionType_USE_HEALTH_POTION,
