+++ Source/engine/render/scrollrt.cpp
@@ -13,6 +13,7 @@
 
 #include "DiabloUI/ui_flags.hpp"
 #include "automap.h"
+#include "controls/control_mode.hpp"
 #include "controls/plrctrls.h"
 #include "cursor.h"
 #include "dead.h"
