+++ Source/track.cpp
@@ -7,6 +7,7 @@
 
 #include <SDL.h>
 
+#include "controls/control_mode.hpp"
 #include "controls/game_controls.h"
 #include "controls/plrctrls.h"
 #include "cursor.h"
